combine mock api calls in activity.ts with scheduler.ts
