+++ src/components/Calendar/CalendarSidebar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button } from '@/components/ui/button';
 import { Label } from '@/components/ui/label';
-import { ActivityTemplate } from '@/api/activity/activity';
+import { ActivityTemplate } from '@/api/scheduler/scheduler';
 
 interface CalendarSidebarProps {
   activityTemplates: ActivityTemplate[];
