+++ src/components/Calendar/views/CentreDailyScheduleView.tsx
@@ -3,7 +3,7 @@ import { format, isSameDay, parseISO, getHours, parse } from "date-fns";
 import {
   ScheduledCentreActivity,
   ActivityTemplate,
-} from "@/api/activity/activity";
+} from "@/api/scheduler/scheduler";
 import { TIME_SLOTS, ACTIVITY_STYLES } from "../CalendarTypes";
 
 interface CentreDailyScheduleViewProps {
